requirements.txt
aiogram==3.18.0
asyncpg==0.30.0
SQLAlchemy==2.0.38
python-dotenv==1.0.1
aiofiles==24.1.0
aiohttp==3.11.13
pydantic==2.10.6